[REF] web, web_editor, website: remove mixins.js
To discourage the use of those mixins outside of their remaining usecase
and as a new steps towards the future refactoring of public widgets, now
move those mixins directly in the public widgets file.

They were also reduced to their strict minimum (features used in the
current Odoo codebase).

Related to task-3930204
