+++ addons/web/static/tests/legacy/public/public_widget_tests.js
@@ -326,76 +326,6 @@ QUnit.module('core', {}, function () {
 
     QUnit.module('Widget, and async stuff');
 
-    QUnit.test("alive(alive)", async function (assert) {
-        assert.expect(1);
-
-        var widget = new (Widget.extend({}));
-
-        await widget.start()
-            .then(function () {return widget.alive(Promise.resolve()) ;})
-            .then(function () { assert.ok(true); });
-
-        widget.destroy();
-    });
-
-    QUnit.test("alive(dead)", function (assert) {
-        assert.expect(1);
-        var widget = new (Widget.extend({}));
-
-        return new Promise(function (resolve, reject) {
-            widget.start()
-            .then(function () {
-                // destroy widget
-                widget.destroy();
-                var promise = Promise.resolve();
-                // leave time for alive() to do its stuff
-                promise.then(function () {
-                    return Promise.resolve();
-                }).then(function () {
-                    assert.ok(true);
-                    resolve();
-                });
-                // ensure that widget.alive() refuses to resolve or reject
-                return widget.alive(promise);
-            }).then(function () {
-                reject();
-                assert.ok(false, "alive() should not terminate by default");
-            }).catch(function() {
-                reject();
-                assert.ok(false, "alive() should not terminate by default");
-            });
-        });
-    });
-
-    QUnit.test("alive(alive, true)", async function (assert) {
-        assert.expect(1);
-        var widget = new (Widget.extend({}));
-        await widget.start()
-            .then(function () { return widget.alive(Promise.resolve(), true); })
-            .then(function () { assert.ok(true); });
-        widget.destroy();
-    });
-
-    QUnit.test("alive(dead, true)", function (assert) {
-        assert.expect(1);
-        var done = assert.async();
-
-        var widget = new (Widget.extend({}));
-
-        widget.start()
-        .then(function () {
-            // destroy widget
-            widget.destroy();
-            return widget.alive(Promise.resolve(), true);
-        }).then(function () {
-            assert.ok(false, "alive(p, true) should fail its promise");
-            done();
-        }, function () {
-            assert.ok(true, "alive(p, true) should fail its promise");
-            done();
-        });
-    });
-
     QUnit.test('start is not called when widget is destroyed', function (assert) {
         assert.expect(0);
         const $fix = $("#qunit-fixture");
