+++ debian/copyright
@@ -334,10 +334,6 @@ Files: addons/web/static/fonts/lato/*
 Copyright: 2010-2011 tyPoland Lukasz Dziedzic
 License: SIL-Open-Font-License
 
-Files: addons/web/static/src/js/core/mixins.js
-Copyright: 2010-2012 Jeremy Ashkenas
-License: MIT
-
 Files: addons/website/static/src/libs/zoomodoo/zoomodoo.js
 Copyright: 2013 Matt Hinchliffe
 License: MIT
